Remove SDS, begin refactoring to move away from CPP.
This refactoring will make the whole system more flexible and make CRC checks
more easily possible. What is missing is the sequence stack and small
refactorings of old code portions.

The main change is the format of blocks in memory. Instead of always storing
them in their on-the-wire format, they are stored in a decoded array. Only when
sending a message, it gets converted into its MIDI-like on-the-wire format.